Reduce unnecessary identity queries (#1)
Added tableIsAutoIncrement parameter to InlineInsert. This is a breaking change to anyone directly using InlineInsert, which should be nobody. Updated code generation to use this new constructor. Modified logic of Execute to do what it already said it was going to do - return null if it's not an auto-increment table. This saves a round trip to the database for tables without an identity primary key.